Add files via upload
Group A (Aliffa, Akhilesh, Clairine, Ivander, Radit, Kris, & Katarina)'s Final Project for Object Oriented Programming Class

This project is a microphone-based language translator with the capability of translating many languages. The microphone detects our language from our voice, translating our input and output based on the languages specified as the input and output and vice versa.

Note: For the first translation phase in [Language] to Hindi, the font wouldn't load on some devices because of the unsupported font. In addition, it will appear as some squares. To be able to see the Hindi word, please copy and paste the translation and paste it on Notepad or Word, thank you.